# src/openapi/spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cantor — API interne",
    "version": "1.0.0",
    "description": "API REST interne de Cantor. Usage : documentation et maintenance, pas un accès public tiers. Authentification par cookie de session Supabase — connectez-vous via /login dans le même navigateur avant d'utiliser \"Try it out\"."
  },
  "servers": [{ "url": "/", "description": "Origine courante" }],
  "security": [{ "cookieAuth": [] }],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sb-access-token",
        "description": "Cookie httpOnly géré par Supabase Auth. Non manipulable depuis Swagger UI directement : connectez-vous via /login dans un autre onglet du même navigateur, le cookie suit automatiquement (requêtes same-origin)."
      }
    },
...
  },
  "paths": {
...
    "/api/chants/{id}/validate": {
      "post": {
        "tags": ["Validation"],
        "summary": "Valider un chant — chef de chœur uniquement (en_attente → validé)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "OK" }, "403": { "description": "Réservé au chef de chœur" } }
      }
    },
    "/api/chants/{id}/reject": {
      "post": {
        "tags": ["Validation"],
        "summary": "Rejeter un chant avec un motif — chef de chœur uniquement",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "note": { "type": "string" } }, "required": ["note"] } } } },
        "responses": { "200": { "description": "OK" }, "400": { "description": "Motif manquant" }, "403": { "description": "Réservé au chef de chœur" } }
      }
    },
...
  }
}